Add stat modifier actions & speedy effect
Call Status Effect Apply/Remove
Change stats to use floats for more precise modifications
Add StatContainer#applyModifier/removeModifier

+++ core/src/main/java/dev/creoii/dungeoneer/util/PacketUtils.java
@@ -174,12 +174,12 @@ public final class PacketUtils {
     }
 
     public static Stat readStat(Input input) {
-        return new Stat(Stat.Type.values()[input.readInt()], input.readInt());
+        return new Stat(Stat.Type.values()[input.readInt()], input.readFloat());
     }
 
     public static void writeStat(Output output, Stat stat) {
         output.writeInt(stat.type().ordinal());
-        output.writeInt(stat.value());
+        output.writeFloat(stat.value());
     }
 
     public static void writeStatContainer(Output output, StatContainer container) {
