+++ core/src/main/java/dev/creoii/dungeoneer/util/stat/ModifierEntry.java
@@ -6,12 +6,12 @@ import dev.creoii.dungeoneer.util.Codecs;
 
 import java.util.UUID;
 
-public record ModifierEntry(Stat.Type type, UUID uuid, int amount, Operation operation, ModifierType modifierType) {
+public record ModifierEntry(Stat.Type type, UUID uuid, float amount, Operation operation, ModifierType modifierType) {
     public static final Codec<ModifierEntry> CODEC = RecordCodecBuilder.create(instance -> {
         return instance.group(
             Stat.Type.CODEC.fieldOf("type").forGetter(ModifierEntry::type),
             Codecs.UUID.fieldOf("uuid").orElse(UUID.randomUUID()).forGetter(ModifierEntry::uuid),
-            Codec.INT.fieldOf("amount").orElse(0).forGetter(ModifierEntry::amount),
+            Codec.FLOAT.fieldOf("amount").orElse(0f).forGetter(ModifierEntry::amount),
             Operation.CODEC.fieldOf("operation").orElse(Operation.NONE).forGetter(ModifierEntry::operation),
             ModifierType.CODEC.fieldOf("modifier_type").orElse(ModifierType.ALL).forGetter(ModifierEntry::modifierType)
         ).apply(instance, ModifierEntry::new);
